edit pep8 violations

+++ pysnips/data_science/wine_data/__init__.py
@@ -32,7 +32,7 @@ class RedWineData(object):
         :rtype: list
         """
         file_name = self.download_red_wine()[0]
-        wines = np.genfromtxt(fname=file_name, delimiter=";", skip_header=1,dtype=np.float)
+        wines = np.genfromtxt(fname=file_name, delimiter=";", skip_header=1)
         # alternatively
         # with open(file_name, "r") as f:
         #     wines = list(csv.reader(f, delimiter=";"))
